freetype: Build with -fno-strict-aliasing
Freetype's code is not quite ready to be built with strict
aliasing rules enforced.
Add -fno-strict-aliasing here so we can disable the global
-fno-strict-aliasing bit.

Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>

freetype: Don't force -O2

There's a -O2 hardcode in Android.mk - its purpose is to make
sure freetype gets built with optimizations enabled, but it
is done in a way that overrides other optimization flags such
as -O3 or -Os.

Only set -O2 if no other optimization level is requested.

Conflicts:

	Android.mk

Change-Id: I1c6191c850afa940a89269a95ce686c2fd422640
Signed-off-by: Bernhard Rosenkraenzer <Bernhard.Rosenkranzer@linaro.org>
Signed-off-by: Evan McClain <aeroevan@gmail.com>